deteksi-web/requirements.txt
Flask==2.3.2
gunicorn==21.2.0
numpy==1.24.3
tensorflow==2.13.0
opencv-python==4.8.0.76
Pillow==10.2.0